// libs/features/collection/billing/src/lib/shared/models/bill-history-wrapper.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

import { BillHistory } from './bill-history';
import { GosiCalendar } from '@gosi-ui/core/lib/models/gosi-calendar';
import { BilingualText } from '@gosi-ui/core';

export class BillHistoryWrapper {
  billBatchIndicator: boolean;
  availableBillStartDate: GosiCalendar[] = [];
  bills: BillHistory[];
  firstBillIssueDate: GosiCalendar = new GosiCalendar();
  firstBillIssueMonth: BilingualText;
  lastBillIssueDate?: GosiCalendar = new GosiCalendar();
  lastFullyPaidDate?: GosiCalendar = new GosiCalendar();
  lastMigratedBillStartDate?: GosiCalendar = new GosiCalendar();
  fullyPaidMonth: null;
  noOfMonthsSinceLastPaid: number;
  totalNoOfRecords: number;
  contributorName: BilingualText;
  installmentComplianceInd: BilingualText;
  requiredMinPayment: number;
  migratedBillPaidAmount: number;
  migratedBillUnPaidAmount: number;
  migratedOutstandingAmount: number;
  noOfPaidContribution: number;
  debtStartDate : GosiCalendar = new GosiCalendar();
}